+++ src/global/misc/src/MResponseMatrixO11.cxx
@@ -839,11 +839,11 @@ float MResponseMatrixO11::GetMinimum() const
 ////////////////////////////////////////////////////////////////////////////////
 
 
-float MResponseMatrixO11::GetSum() const
+double MResponseMatrixO11::GetSum() const
 {
   // Return the sum of all bins:
 
-  float Sum = 0;
+  double Sum = 0;
   for (unsigned int i = 0; i < m_AxesO10.size(); ++i) {
     Sum += m_AxesO10[i].GetSum();
   }
