+++ src/global/misc/src/MResponseMatrixO14.cxx
@@ -878,11 +878,11 @@ float MResponseMatrixO14::GetMinimum() const
 ////////////////////////////////////////////////////////////////////////////////
 
 
-float MResponseMatrixO14::GetSum() const
+double MResponseMatrixO14::GetSum() const
 {
   // Return the sum of all bins:
 
-  float Sum = 0;
+  double Sum = 0;
   for (unsigned int i = 0; i < m_AxesO13.size(); ++i) {
     Sum += m_AxesO13[i].GetSum();
   }
